+++ docs/changelog.md
@@ -11,6 +11,7 @@ and version numbering adheres to [Semantic Versioning](http://semver.org/spec/v2
   - [Added]   first `--cleanup` implementation, that removes files created during the test by bbt
   - [Added]   in scenarios, `dir` is now a synonym of `directory`
   - [Changed] bbt bootstraps! bbt now runs bbt tests that return error code.
+  - [Added]   `--exec_dir` option to run scenario in a dir different from current
   
 - [0.0.3]
   - [Added] text file creation